connector plus driver merge
