+++ runtime/compilers/rillv1beta/connector.go
@@ -13,6 +13,7 @@ import (
 	"github.com/rilldata/rill/runtime/services/catalog/artifacts"
 	"github.com/rilldata/rill/runtime/services/catalog/migrator/models"
 	"github.com/rilldata/rill/runtime/services/catalog/migrator/sources"
+	"go.uber.org/zap"
 )
 
 // TODO :: return this to build support for all kind of variables
@@ -70,7 +71,7 @@ func ExtractConnectors(ctx context.Context, projectPath string) ([]*Connector, e
 		}
 		// ignoring error since failure to resolve this should not break the deployment flow
 		// this can fail under cases such as full or host/bucket of URI is a variable
-		access, _ := connector.HasAnonymousAccess(ctx, src.Properties.AsMap())
+		access, _ := connector.HasAnonymousSourceAccess(ctx, source(src.Connector, src), zap.NewNop())
 		c := key{Name: src.Connector, Type: src.Connector, AnonymousAccess: access}
 		srcs, ok := connectorMap[c]
 		if !ok {
@@ -156,4 +157,30 @@ type key struct {
 	Name            string
 	Type            string
 	AnonymousAccess bool
+}
+
+func source(connector string, src *runtimev1.Source) drivers.Source {
+	props := src.Properties.AsMap()
+	switch connector {
+	case "s3":
+		return &drivers.BucketSource{
+			Properties: props,
+		}
+	case "gcs":
+		return &drivers.BucketSource{
+			Properties: props,
+		}
+	case "https":
+		return &drivers.FileSource{
+			Properties: props,
+		}
+	case "local_file":
+		return &drivers.FileSource{
+			Properties: props,
+		}
+	case "motherduck":
+		return &drivers.DatabaseSource{}
+	default:
+		return nil
+	}
 }
