--- runtime/drivers/gcs/connectors.go
@@ -1,97 +1,0 @@
-package gcs
-
-import (
-	"context"
-	"fmt"
-	"os"
-	"strings"
-
-	"github.com/rilldata/rill/runtime/drivers"
-	"github.com/rilldata/rill/runtime/pkg/fileutil"
-	"gocloud.dev/blob/gcsblob"
-	"gocloud.dev/gcp"
-)
-
-func init() {
-	drivers.RegisterConnector("gcs", &Connection{})
-}
-
-var spec = drivers.Spec{
-	DisplayName:        "Google Cloud Storage",
-	Description:        "Connect to Google Cloud Storage.",
-	ServiceAccountDocs: "https://docs.rilldata.com/deploy/credentials/gcs",
-	SourceProperties: []drivers.PropertySchema{
-		{
-			Key:         "path",
-			DisplayName: "GS URI",
-			Description: "Path to file on the disk.",
-			Placeholder: "gs://bucket-name/path/to/file.csv",
-			Type:        drivers.StringPropertyType,
-			Required:    true,
-			Hint:        "Glob patterns are supported",
-		},
-		{
-			Key:         "gcp.credentials",
-			DisplayName: "GCP credentials",
-			Description: "GCP credentials inferred from your local environment.",
-			Type:        drivers.InformationalPropertyType,
-			Hint:        "Set your local credentials: <code>gcloud auth application-default login</code> Click to learn more.",
-			Href:        "https://docs.rilldata.com/develop/import-data#configure-credentials-for-gcs",
-		},
-	},
-	ConfigProperties: []drivers.PropertySchema{
-		{
-			Key:  "google_application_credentials",
-			Hint: "Enter path of file to load from.",
-			ValidateFunc: func(any interface{}) error {
-				val := any.(string)
-				if val == "" {
-					// user can chhose to leave empty for public sources
-					return nil
-				}
-
-				path, err := fileutil.ExpandHome(strings.TrimSpace(val))
-				if err != nil {
-					return err
-				}
-
-				_, err = os.Stat(path)
-				return err
-			},
-			TransformFunc: func(any interface{}) interface{} {
-				val := any.(string)
-				if val == "" {
-					return ""
-				}
-
-				path, err := fileutil.ExpandHome(strings.TrimSpace(val))
-				if err != nil {
-					return err
-				}
-				// ignoring error since PathError is already validated
-				content, _ := os.ReadFile(path)
-				return string(content)
-			},
-		},
-	},
-}
-
-func (c *Connection) HasAnonymousAccess(ctx context.Context, props map[string]any) (bool, error) {
-	conf, err := parseConfig(props)
-	if err != nil {
-		return false, fmt.Errorf("failed to parse config: %w", err)
-	}
-
-	client := gcp.NewAnonymousHTTPClient(gcp.DefaultTransport())
-	bucketObj, err := gcsblob.OpenBucket(ctx, client, conf.url.Host, nil)
-	if err != nil {
-		return false, fmt.Errorf("failed to open bucket %q, %w", conf.url.Host, err)
-	}
-
-	return bucketObj.IsAccessible(ctx)
-}
-
-// ConnectorSpec implements drivers.Connection.
-func (c *Connection) Spec() drivers.Spec {
-	return spec
-}
