+++ runtime/drivers/gcs/gcs.go
@@ -5,12 +5,14 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
+	"os"
 	"strings"
 
 	"github.com/bmatcuk/doublestar/v4"
 	"github.com/mitchellh/mapstructure"
 	"github.com/rilldata/rill/runtime/drivers"
 	rillblob "github.com/rilldata/rill/runtime/drivers/blob"
+	"github.com/rilldata/rill/runtime/pkg/fileutil"
 	"github.com/rilldata/rill/runtime/pkg/globutil"
 	"go.uber.org/zap"
 	"gocloud.dev/blob/gcsblob"
@@ -26,13 +28,85 @@ var errNoCredentials = errors.New("empty credentials: set `google_application_cr
 
 func init() {
 	drivers.Register("gcs", driver{})
+	drivers.RegisterAsConnector("gcs", driver{})
+}
+
+var spec = drivers.Spec{
+	DisplayName:        "Google Cloud Storage",
+	Description:        "Connect to Google Cloud Storage.",
+	ServiceAccountDocs: "https://docs.rilldata.com/deploy/credentials/gcs",
+	SourceProperties: []drivers.PropertySchema{
+		{
+			Key:         "path",
+			DisplayName: "GS URI",
+			Description: "Path to file on the disk.",
+			Placeholder: "gs://bucket-name/path/to/file.csv",
+			Type:        drivers.StringPropertyType,
+			Required:    true,
+			Hint:        "Glob patterns are supported",
+		},
+		{
+			Key:         "gcp.credentials",
+			DisplayName: "GCP credentials",
+			Description: "GCP credentials inferred from your local environment.",
+			Type:        drivers.InformationalPropertyType,
+			Hint:        "Set your local credentials: <code>gcloud auth application-default login</code> Click to learn more.",
+			Href:        "https://docs.rilldata.com/develop/import-data#configure-credentials-for-gcs",
+		},
+	},
+	ConfigProperties: []drivers.PropertySchema{
+		{
+			Key:  "google_application_credentials",
+			Hint: "Enter path of file to load from.",
+			ValidateFunc: func(any interface{}) error {
+				val := any.(string)
+				if val == "" {
+					// user can chhose to leave empty for public sources
+					return nil
+				}
+
+				path, err := fileutil.ExpandHome(strings.TrimSpace(val))
+				if err != nil {
+					return err
+				}
+
+				_, err = os.Stat(path)
+				return err
+			},
+			TransformFunc: func(any interface{}) interface{} {
+				val := any.(string)
+				if val == "" {
+					return ""
+				}
+
+				path, err := fileutil.ExpandHome(strings.TrimSpace(val))
+				if err != nil {
+					return err
+				}
+				// ignoring error since PathError is already validated
+				content, _ := os.ReadFile(path)
+				return string(content)
+			},
+		},
+	},
 }
 
 type driver struct{}
 
+type configProperties struct {
+	SecretJSON      string `mapstructure:"google_application_credentials"`
+	AllowHostAccess bool   `mapstructure:"allow_host_access"`
+}
+
 func (d driver) Open(config map[string]any, logger *zap.Logger) (drivers.Connection, error) {
+	conf := &configProperties{}
+	err := mapstructure.Decode(config, conf)
+	if err != nil {
+		return nil, err
+	}
+
 	conn := &Connection{
-		config: config,
+		config: conf,
 		logger: logger,
 	}
 	return conn, nil
@@ -42,7 +116,31 @@ func (d driver) Drop(config map[string]any, logger *zap.Logger) error {
 	return drivers.ErrDropNotSupported
 }
 
-type config struct {
+func (d driver) Spec() drivers.Spec {
+	return spec
+}
+
+func (d driver) HasAnonymousSourceAccess(ctx context.Context, src drivers.Source, logger *zap.Logger) (bool, error) {
+	b, ok := src.BucketSource()
+	if !ok {
+		return false, fmt.Errorf("require bucket source")
+	}
+
+	conf, err := parseSourceProperties(b.Properties)
+	if err != nil {
+		return false, fmt.Errorf("failed to parse config: %w", err)
+	}
+
+	client := gcp.NewAnonymousHTTPClient(gcp.DefaultTransport())
+	bucketObj, err := gcsblob.OpenBucket(ctx, client, conf.url.Host, nil)
+	if err != nil {
+		return false, fmt.Errorf("failed to open bucket %q, %w", conf.url.Host, err)
+	}
+
+	return bucketObj.IsAccessible(ctx)
+}
+
+type sourceProperties struct {
 	Path                  string `key:"path"`
 	GlobMaxTotalSize      int64  `mapstructure:"glob.max_total_size"`
 	GlobMaxObjectsMatched int    `mapstructure:"glob.max_objects_matched"`
@@ -51,8 +149,8 @@ type config struct {
 	url                   *globutil.URL
 }
 
-func parseConfig(props map[string]any) (*config, error) {
-	conf := &config{}
+func parseSourceProperties(props map[string]any) (*sourceProperties, error) {
+	conf := &sourceProperties{}
 	err := mapstructure.Decode(props, conf)
 	if err != nil {
 		return nil, err
@@ -76,8 +174,7 @@ func parseConfig(props map[string]any) (*config, error) {
 }
 
 type Connection struct {
-	// config holds google_application_credentials and allow_host_access
-	config map[string]any
+	config *configProperties
 	logger *zap.Logger
 }
 
@@ -90,7 +187,9 @@ func (c *Connection) Driver() string {
 
 // Config implements drivers.Connection.
 func (c *Connection) Config() map[string]any {
-	return c.config
+	m := make(map[string]any, 0)
+	_ = mapstructure.Decode(c.config, m)
+	return m
 }
 
 // Close implements drivers.Connection.
@@ -147,7 +246,7 @@ func (c *Connection) AsFileStore() (drivers.FileStore, bool) {
 // The credential json is read from config google_application_credentials.
 // Additionally in case `allow_host_credentials` is true it looks for "Application Default Credentials" as well
 func (c *Connection) DownloadFiles(ctx context.Context, source *drivers.BucketSource) (drivers.FileIterator, error) {
-	conf, err := parseConfig(source.Properties)
+	conf, err := parseSourceProperties(source.Properties)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse config: %w", err)
 	}
@@ -206,16 +305,12 @@ func (c *Connection) createClient(ctx context.Context) (*gcp.HTTPClient, error) 
 }
 
 func (c *Connection) resolvedCredentials(ctx context.Context) (*google.Credentials, error) {
-	if secretJSON := c.config["google_application_credentials"].(string); secretJSON != "" {
+	if c.config.SecretJSON != "" {
 		// google_application_credentials is set, use credentials from json string provided by user
-		return google.CredentialsFromJSON(ctx, []byte(secretJSON), "https://www.googleapis.com/auth/cloud-platform")
+		return google.CredentialsFromJSON(ctx, []byte(c.config.SecretJSON), "https://www.googleapis.com/auth/cloud-platform")
 	}
 	// google_application_credentials is not set
-	allowHostAccess := false
-	if val, ok := c.config["allow_host_access"].(bool); ok {
-		allowHostAccess = val
-	}
-	if allowHostAccess {
+	if c.config.AllowHostAccess {
 		// use host credentials
 		creds, err := gcp.DefaultCredentials(ctx)
 		if err != nil {
